<type>(<scope>): <subject> <BLANK LINE> <body> <BLANK LINE> <footer>

+++ common/common-log/src/main/java/com/feidian/aspect/CommodityOperLogAspect.java
@@ -54,7 +54,7 @@ public class CommodityOperLogAspect {
 
             //请求的用户
             String token = request.getHeader("token");
-            Integer userId = JwtUtil.getSubject(token);
+            long userId = JwtUtil.getSubject(token);
             User user = commodityOperLogService.findById(userId);
             commodityOperLog.setOperUserName(user.getUsername());
